feat: update the animal game screen controller (notifier) class
